feat(plan): /plan 단일 통합 화면 신설 + ItemPanel 데이터 연결 + 중복 정리
Phase D - /plan 단일 통합:
- PlanScreen 컴포넌트 추출 (URL 기반 ?item=&day= 동기)
- /plan 라우트 신설 (메인 작업 화면)
- /map → /plan 으로 redirect (북마크 호환, search params 유지)
- / (홈) → /plan 으로 redirect
- Navigation 메뉴: 계획(/plan) / 목록(/research) / 일정(/schedule)
- /research 의 ViewToggle '일정' 모드 제거 (목록 전용으로 단순화).
  '일정' 메뉴와 중복되던 부분 해소

Phase E - ItemPanel onSave/onDelete 결선:
- PlanScreen·/research·/schedule 모두 onSave 시 success toast,
  onDelete 시 selection cleanup + success toast
- ItemPanel 내부 useItems.updateItem/deleteItem 은 이미 작동 중,
  호스트 콜백을 toast/cleanup 으로 명시 결선

기타:
- 모바일 ThemeToggle 접근성: /research·/schedule 헤더 우측, /plan 우측 상단 fixed
- FAB 위치를 className override 가능하게 변경
