+++ app/plan/page.tsx
@@ -1,0 +1,12 @@
+'use client'
+
+import { Suspense } from 'react'
+import PlanScreen from '@/components/Plan/PlanScreen'
+
+export default function PlanPage() {
+  return (
+    <Suspense fallback={null}>
+      <PlanScreen basePath="/plan" />
+    </Suspense>
+  )
+}
